Projects uploaded
Making sense out of a dataset is one of the most fundamental skills of a computer scientist.  These assignment's do just that according to the widely accepted principals of normal forms.  I was told to use draw.io for the visuals from a mentor, but I use paint.net more proficiently now a days in my web development work.